### 2020-12-16 #2915 @eribul
This package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/381, @ropensci/editors
