### 2020-12-16 ropensci/coder#143 @eribul
Keep track of JOSS
https://github.com/openjournals/joss-reviews/issues/2915
