### 2020-12-16 #2915 @kthyng
HI @eribul! I see that this has already been reviewed with rOpenSci. Because of that, I can take the submission through the whole JOSS process.
