### 2020-12-16 #2916 @whedon
[REVIEW]: coder: An R package for code-based item classification and categorization
**Submitting author:** @eribul (<a href="http://orcid.org/0000-0002-9973-456X">Erik Bülow</a>)
**Repository:** <a href="https://github.com/ropensci/coder" target ="_blank">https://github.com/ropensci/coder</a>
**Version:** v0.13.2
