### 2020-12-16 #2916 @kthyng
This submission was previously reviewed with rOpenSci: https://github.com/ropensci/software-review/issues/381
