### 2020-12-16 #2916 @kthyng
@eribul Looking through your paper, it looks good but the references need a bit of attention. For example:

* Gasparini has "r" lower case (you can force capitalization with {} around one or more characters)
* Jawad et al seems to have a typo in the title
* the author list for Quan et al has lower case letters for initials
* Pratt et al has "rx-risk", not sure if any of that should be capitalized.

Could you go through the references in detail and make sure everything is correct?
