### 2020-12-16 #2916 @kthyng
@eribul Also can you archive your repo with a place like Zenodo and report the doi back here? Be sure to edit the metadata of the archive so that the title and author list exactly match that of your JOSS paper.
