### 2020-12-16 #2916 @eribul
Dear @kthyng thank you very much for very rapid response and nice feedback!
I have now included a section called "Statement of Need" (changed one of the old headings).
I have changed the references according to your suggestions.
I also found some additional typos in the references, which are also fixed.
I have a repo at Zenodo: http://doi.org/10.5281/zenodo.4329526 with the same title and author as for the paper.
