### 2020-12-17 #2916 @kthyng
@eribul Excellent, things looks good but I did notice one more thing with your paper. Could you capitalize the appropriate items in your second affiliation? It currently looks a bit informal.

What software version would you like associated with your submission?
