### 2020-12-18 #2916 @kthyng
Congrats on your JOSS publication @eribul!!
